API security update

+++ app/backend/src/main/java/com/wishlist/controllers/FamilyController.java
@@ -5,6 +5,9 @@ import com.wishlist.dto.CreateFamilyRequestDTO;
 import com.wishlist.dto.FamilyMemberDTO;
 import com.wishlist.exceptions.FamilyDoesNotExistException;
 import com.wishlist.models.Family;
+import com.wishlist.models.User;
+import com.wishlist.security.JwtGeneratorImpl;
+import com.wishlist.security.JwtValidator;
 import com.wishlist.services.FamilyService;
 import com.wishlist.services.UserService;
 import org.slf4j.Logger;
@@ -22,23 +25,32 @@ public class FamilyController {
     Logger logger = LoggerFactory.getLogger(FamilyController.class);
     private final FamilyService familyService;
     private final UserService userService;
+    private final JwtValidator jwtValidator;
 
-    public FamilyController(final FamilyService familyService, final UserService userService) {
+    public FamilyController(final FamilyService familyService, final UserService userService, final JwtValidator jwtValidator) {
         this.familyService = familyService;
         this.userService = userService;
+        this.jwtValidator = jwtValidator;
     }
 
-    @GetMapping
+/*    @GetMapping
     public List<Family> getAll() {
-        return familyService.getAll();
-    }
+        List<Family> families = familyService.getAll();
+        return families;
+    }*/
 
     @GetMapping("/{id}")
-    public ResponseEntity<?> getFamily(@PathVariable String id) {
-        logger.info("get fml {}", id);
+    public ResponseEntity<?> getFamily(@PathVariable String id, @RequestHeader("Authorization") String jwt) {
         try {
             logger.info("rtn fml {}", id);
-            return new ResponseEntity(familyService.findById(id), HttpStatus.OK);
+
+            if(jwtValidator.validateFamily(jwt, id)){
+                return new ResponseEntity(familyService.findById(id), HttpStatus.OK);
+            }
+            else{
+                logger.info("rtn fml {} FAIL", id);
+                return new ResponseEntity(new ApiError("user does not have access to this family"), HttpStatus.UNAUTHORIZED);
+            }
         } catch (FamilyDoesNotExistException e) {
             logger.info("rtn fml {} FAIL", id);
             return new ResponseEntity(new ApiError(e.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
@@ -46,54 +58,71 @@ public class FamilyController {
     }
 
     @GetMapping("/{id}/members")
-    public ResponseEntity<?> getFamilyMembers(@PathVariable String id) {
-        logger.info("get fml mbr for {}",id);
+    public ResponseEntity<?> getFamilyMembers(@PathVariable String id, @RequestHeader("Authorization") String jwt) {
         try {
-            logger.info("get fml mbr for {} success",id);
-            return new ResponseEntity<>(familyService.getFamilyMembers(id).stream().map(FamilyMemberDTO::to).collect(Collectors.toList()), HttpStatus.OK);
+            if(jwtValidator.validateFamily(jwt, id)){
+                return new ResponseEntity<>(familyService.getFamilyMembers(id).stream().map(FamilyMemberDTO::to).collect(Collectors.toList()), HttpStatus.OK);
+            } else {
+                return new ResponseEntity(new ApiError("User does not have access to this family"), HttpStatus.UNAUTHORIZED);
+            }
         } catch (FamilyDoesNotExistException e) {
-            logger.info("get fml mbr for {} fail",id);
             return new ResponseEntity<>(new ApiError(e.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @GetMapping("/user/{userId}")
-    public ResponseEntity<?> getFamilyForUser(@PathVariable String userId) {
+    public ResponseEntity<?> getFamilyForUser(@PathVariable String userId, @RequestHeader("Authorization") String jwt) {
         logger.info("get fml for usr {}", userId);
         try {
             logger.info("rtn fml for usr {}", userId);
-            return new ResponseEntity<>(familyService.findByUser(userId), HttpStatus.OK);
+            if(jwtValidator.validateUser(jwt, userId)){
+                return new ResponseEntity<>(familyService.findByUser(userId), HttpStatus.OK);
+            }
+            else{
+                return new ResponseEntity<>(new ApiError("User does not have access to this family"), HttpStatus.INTERNAL_SERVER_ERROR);
+            }
         } catch (FamilyDoesNotExistException e) {
             logger.info("rtn fml for usr {} FAIL", userId);
             return new ResponseEntity<>(new ApiError(e.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
-    @DeleteMapping("/{id}")
-    public ResponseEntity<?> delete(@PathVariable String id) {
+    @DeleteMapping("/{familyId}")
+    public ResponseEntity<?> delete(@PathVariable String familyId, @RequestHeader("Authorization") String jwt) {
         try {
-            return new ResponseEntity<>(familyService.delete(id), HttpStatus.OK);
+
+            if (jwtValidator.validateFamily(jwt, familyId) && familyService.isOwner(jwtValidator.getUserFromJwt(jwt).getId(), familyId)){
+                return new ResponseEntity<>(familyService.delete(familyId), HttpStatus.OK);
+            }
+            else{
+                return new ResponseEntity<>(new ApiError("You do not have access to this user"), HttpStatus.INTERNAL_SERVER_ERROR);
+            }
         } catch (FamilyDoesNotExistException e) {
             return new ResponseEntity<>(new ApiError(e.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @PostMapping("/{userId}")
-    public ResponseEntity<?> create(@PathVariable String userId, @RequestBody CreateFamilyRequestDTO dto) {
+    public ResponseEntity<?> create(@PathVariable String userId, @RequestBody CreateFamilyRequestDTO dto, @RequestHeader("Authorization") String jwt) {
         logger.info("POST createFamily");
         try {
-            logger.info("POST createFamily fml crt HTTP 200");
-            Family saved = familyService.save(CreateFamilyRequestDTO.from(dto));
-            userService.addUserToFamily(userId, saved.getId());
-            return new ResponseEntity<>(saved, HttpStatus.CREATED);
+            if (jwtValidator.validateUser(jwt, userId)){
+                Family saved = familyService.saveWithOwner(CreateFamilyRequestDTO.from(dto), userService.getUserById(userId));
+                userService.addUserToFamily(userId, saved.getId());
+                logger.info("POST createFamily fml crt HTTP 200");
+                return new ResponseEntity<>(saved, HttpStatus.CREATED);
+            }
+            else{
+                return new ResponseEntity<>(new ApiError("You do not have access to this user"), HttpStatus.INTERNAL_SERVER_ERROR);
+            }
         } catch (Exception e) {
             logger.error("POST createFamily FAIL {}", e.getMessage());
             return new ResponseEntity<>(new ApiError(e.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
 
         }
     }
-
-    @PutMapping("/{id}")
+    //i don't like this route
+/*    @PutMapping("/{id}")
     public ResponseEntity<Family> update(@PathVariable String id, @RequestBody Family updatedFamily) {
         //todo move to service layer
         try {
@@ -105,12 +134,32 @@ public class FamilyController {
             return ResponseEntity.notFound().build();
 
         }
+    }*/
+
+    //how about we use this route instead?
+    @PostMapping("/remove/{familyId}/{userId}")
+    public ResponseEntity<Family> removeFromFamily(@PathVariable("familyId") String familyId, @PathVariable("userId") String userId, @RequestHeader("Authorization") String jwt){
+        try {
+            if (jwtValidator.validateUser(jwt, userId) && familyService.isOwner(familyId, userId)){
+                return new ResponseEntity<>(familyService.removeUserFromFamily(familyId, userId), HttpStatus.OK);
+            }
+            else {
+                return new ResponseEntity(new ApiError("User does not have access to this family"), HttpStatus.INTERNAL_SERVER_ERROR);
+            }
+        } catch (Exception e) {
+            return new ResponseEntity(new ApiError(e.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @PostMapping("/leave/{familyId}/{userId}")
     public ResponseEntity<Family> leaveFamily(@PathVariable("familyId") String familyId, @PathVariable("userId") String userId) {
         try {
-            return new ResponseEntity<>(familyService.removeUserFromFamily(familyId, userId), HttpStatus.OK);
+            if (!familyService.isOwner(familyId, userId)){
+                return new ResponseEntity<>(familyService.removeUserFromFamily(familyId, userId), HttpStatus.OK);
+            }
+            else {
+                return new ResponseEntity(new ApiError("owner cannot leave family"), HttpStatus.INTERNAL_SERVER_ERROR);
+            }
         } catch (Exception e) {
             return new ResponseEntity(new ApiError(e.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
         }
