+++ app/backend/src/main/java/com/wishlist/controllers/InvitationController.java
@@ -1,6 +1,7 @@
 package com.wishlist.controllers;
 
 import com.wishlist.models.Invitation;
+import com.wishlist.security.JwtValidator;
 import com.wishlist.services.interfaces.IInvitationService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,33 +14,55 @@ import java.util.List;
 public class InvitationController {
 
     private final IInvitationService invitationService;
+    private final JwtValidator jwtValidator;
 
-    public InvitationController(IInvitationService invitationService) {
+    public InvitationController(final IInvitationService invitationService, final JwtValidator jwtValidator) {
         this.invitationService = invitationService;
+        this.jwtValidator = jwtValidator;
     }
 
-    @GetMapping
+/*    @GetMapping
     public List<Invitation> getAll() {
         return invitationService.getAllInvitations();
-    }
+    }*/
 
     @GetMapping("/{id}")
-    public ResponseEntity<Invitation> getInvitation(@PathVariable String id) {
-        Invitation invitation = invitationService.findById(id);
-        if (invitation!=null) {
-            return new ResponseEntity<>(invitation, HttpStatus.OK);
-        } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    public ResponseEntity<Invitation> getInvitation(@PathVariable String id, @RequestHeader("Authorization") String jwt) {
+        try{
+            if(jwtValidator.validateInvitation(jwt, id)){
+                Invitation invitation = invitationService.findById(id);
+                if (invitation!=null) {
+                    return new ResponseEntity<>(invitation, HttpStatus.OK);
+                } else {
+                    return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+                }
+            }
+            else{
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            }
+        }
+        catch (Exception e){
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable String id) {
-        boolean success = invitationService.delete(id);
-        if (success) {
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-        } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    public ResponseEntity<Void> delete(@PathVariable String id, @RequestHeader("Authorization") String jwt) {
+        try{
+            if(jwtValidator.validateInvitation(jwt, id)){
+                boolean success = invitationService.delete(id);
+                if (success) {
+                    return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+                } else {
+                    return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+                }
+            }
+            else{
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            }
+        }
+        catch (Exception e){
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
